Move resumable event handlers
